requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
python-multipart>=0.0.5
requests>=2.26.0
beautifulsoup4>=4.9.3
pytube>=12.1.0
pydub>=0.25.1
openai>=0.27.0
python-dotenv>=0.19.0
edge-tts>=6.1.5
moviepy>=1.0.3
soundfile>=0.12.1
numpy>=1.26.0
librosa>=0.10.1
